requirements.txt
numpy
torch
tensorflow
transformers
tensorflow
mtranslate
sentence-transformers
datasets
faiss-cpu
streamlit